Fixed UpdateCustomer issues.

+++ src/controller/CustomerList.java
@@ -1,6 +1,8 @@
 package controller;
 
+import dao.AppointmentDAO;
 import dao.CustomerDAO;
+import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXMLLoader;
@@ -8,12 +10,12 @@ import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
+import model.Appointment;
 import model.Customer;
 import utilities.SceneMovements;
 
 import java.io.IOException;
 import java.net.URL;
-import java.util.Objects;
 import java.util.ResourceBundle;
 
 public class CustomerList implements Initializable{
@@ -42,7 +44,7 @@ public class CustomerList implements Initializable{
 
     }
 
-    public void toUpdateCustomer(ActionEvent actionEvent){
+    public void toUpdateCustomer(ActionEvent actionEvent) throws IOException {
         Customer c = customerListTable.getSelectionModel().getSelectedItem();
         UpdateCustomer.passingTheCustomer(c);
 
@@ -55,7 +57,7 @@ public class CustomerList implements Initializable{
         SceneMovements.goToUpdateCustomer(actionEvent);
     }
 
-    public void toNewCustomer(ActionEvent actionEvent) {
+    public void toNewCustomer(ActionEvent actionEvent) throws IOException {
         SceneMovements.goToNewCustomer(actionEvent);
     }
 
@@ -66,22 +68,31 @@ public class CustomerList implements Initializable{
             alert.setTitle("Select Customer");
             alert.showAndWait();
         }
+        if (AppointmentDAO.checkForCustomerAppointmentLinks(c.getCustomerId()) != 0) {
+            Alert alert = new Alert(Alert.AlertType.ERROR, "Customer cannot be deleted. Remove all appointments associated with this customer.");
+            alert.setTitle("Error");
+            alert.showAndWait();
+        } else {
+            Alert alert2 = new Alert(Alert.AlertType.CONFIRMATION, "Are you sure you want to delete " + c.getCustomerName() + "?",ButtonType.YES,ButtonType.NO);
+            alert2.setTitle("Verify Deletion");
+            alert2.showAndWait().ifPresent(response -> {
+                        if (response == ButtonType.YES) {
+                            CustomerDAO.deleteCustomer(c.getCustomerId());
+                            customerListTable.setItems(CustomerDAO.getAllCustomer());
+                        }
+                    });
+
         //DONE: Complete delete customer method
         //TODO: add reference to lambda
-        Alert alert2 = new Alert(Alert.AlertType.CONFIRMATION, "Are you sure you want to delete " + c.getCustomerName() + "?",ButtonType.YES,ButtonType.NO);
-        alert2.setTitle("Verify Deletion");
-        alert2.showAndWait().ifPresent(response -> {
-            if (response == ButtonType.YES) {
-                CustomerDAO.deleteCustomer(c.getCustomerId());
-            }
-        });
-        //TODO: Refresh table after deletion
+
+        }
+        //DONE: Refresh table after deletion
     }
-    public void onActionToAppt(ActionEvent actionEvent) {
+    public void onActionToAppt(ActionEvent actionEvent) throws IOException {
         SceneMovements.goToApptList(actionEvent);
     }
 
-    public void onActionToReport(ActionEvent actionEvent) {
+    public void onActionToReport(ActionEvent actionEvent) throws IOException {
         SceneMovements.goToReportPage(actionEvent);
     }
